+++ test/src/cognitect/aws/api_test.clj
@@ -1,9 +1,10 @@
 (ns cognitect.aws.api-test
-  (:require [clojure.test :as t :refer [deftest testing is]]
+  (:require [clojure.datafy :as datafy]
+            [clojure.test :as t :refer [deftest is testing]]
             [cognitect.aws.client :as client]
-            [cognitect.aws.http :as http]
             [cognitect.aws.client.api :as aws]
-            [cognitect.aws.client.shared :as shared]))
+            [cognitect.aws.client.shared :as shared]
+            [cognitect.aws.http :as http]))
 
 (deftest test-underlying-http-client
   (testing "defaults to shared client"
@@ -11,6 +12,16 @@
       (is (= #{(shared/http-client)}
              (into #{(shared/http-client)}
                    (->> clients (map (fn [c] (-> c client/-get-info :http-client))))))))))
+
+(deftest test-datafy
+  (let [client (aws/client {:api :s3})
+        data (datafy/datafy client)]
+    (is (map? (:service data)))
+    (is (map? (:metadata (:service data))))
+    (is (map? (:endpoint data)))
+    (is (string? (:region data)))
+    (is (map? (:ops data)))
+    (is (map? (:endpoint data)))))
 
 (deftest test-stop
   (let [call-count (atom 0)